+++ bidscoin/mapper.py
@@ -26,7 +26,11 @@ def createmap(recording: Modules.baseModule,
               template,
               bidsmap_unk) -> None:
 
-    plugins.RunPlugin("SequenceEP", recording)
+    if plugins.RunPlugin("SequenceEP", recording) < 0:
+        logger.warning("Sequence {} discarded by {}"
+                       .format(recording.recIdentity(False), 
+                               "SequenceEP"))
+        return
 
     logger.info("Processing: sub '{}', ses '{}', {} ({} files)"
                 .format(recording.subId(),
@@ -36,7 +40,11 @@ def createmap(recording: Modules.baseModule,
 
     recording.index = -1
     while recording.loadNextFile():
-        plugins.RunPlugin("RecordingEP", recording)
+        if plugins.RunPlugin("RecordingEP", recording) < 0:
+            logger.warning("Recording {} discarded by {}"
+                           .format(recording.recIdentity(), 
+                                   "RecordingEP"))
+            continue
         # checking in the current map
         modality, r_index, r_obj = bidsmap.match_run(recording)
         if not modality:
@@ -243,7 +251,10 @@ def mapper(source: str, destination: str,
         scan = BidsSession()
         scan.in_path = sub_dir
         scan.subject = sub_id
-        plugins.RunPlugin("SubjectEP", scan)
+        if plugins.RunPlugin("SubjectEP", scan) < 0:
+            logger.warning("Subject {} discarded by {}"
+                           .format(scan.subject, "SubjectEP"))
+            continue
         scan.lock_subject()
         if not scan.isSubValid():
             logger.error("{}: Subject id '{}' is not valid"
@@ -272,7 +283,10 @@ def mapper(source: str, destination: str,
                                 ses_dir))
             scan.unlock_session()
             scan.session = os.path.basename(ses_dir)
-            plugins.RunPlugin("SessionEP", scan)
+            if plugins.RunPlugin("SessionEP", scan) < 0:
+                logger.warning("Session {} discarded by {}"
+                               .format(scan.session, "SessionEP"))
+                continue
             scan.lock()
 
             if ses_skip_dir and tools.skipEntity(scan.session,