+++ CHANGELOG.md
@@ -23,7 +23,8 @@ in plug-ins
 - `participants.tsv` file now created at prepearing stage, with warning
 if field values are conflicting
 - Main code is `bidscoin.py` with cli command `prepare`, `bidsify` and `map`
-- Plugins exit code `< 0` do not throw exception  
+- Plugins `SubjectEP`, `SessionEP`, `SequenceEP` and `RecordingEP` can exit with
+code `< 0` will not throw exception but skip corresponding loop
 
 ## [dev2.1.0] - 2020-02-19